requirements.txt
flask==3.0.0
flask_sqlalchemy==3.1.1
flask_migrate==4.0.4
flask_jwt_extended==4.4.4
flask_cors==4.0.0
apscheduler==3.10.1
openai==1.3.3
pytest==8.1.1
pytest-flask==1.2.0
python-dotenv==1.0.1
requests==2.31.0